backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
psycopg[binary]
requests
beautifulsoup4
python-dotenv
